Improve handling of multiple input files
The output filename was set to 'report.html' by default, which is a
problem when generating report for multiple input files at once.
Because the output filename did not change, the HTML reports overwrite
eachother.

The solution is to match the output to the input filename. So an input
of foo.json would result in foo.html output. Thankfully gulp-util's
`replaceFileExtension()` method makes this very easy.

Along the way I changed the creation of the paths to use Node's own
`path` library.